requirements.txt
asgiref==3.8.1
Django==5.0.7
django-ckeditor-5==0.2.13
django-js-asset==2.2.0
django-mptt==0.16.0
pillow==10.4.0
sqlparse==0.5.1
